@@ -1,8 +1,7 @@
-<script>
+<script lang="ts">
   import Banner from "$lib/util/Banner.svelte";
   import Head from "$lib/components/Head.svelte";
-  export let title;
-  export let date;
+  let { title, date } = $props();
 </script>
 
 <Head {title} />
