@@ -1,5 +1,5 @@
-<script>
-  export let summary;
+<script lang="ts">
+  let { summary } = $props();
 </script>
 
 <p>This is the summary from the YAML metadata. It's being served from a component.</p>
