@@ -1,21 +1,27 @@
 <script lang="ts">
+  import { run } from 'svelte/legacy';
+
   import { onMount } from "svelte";
 
-  export let value = 10;
-  export let delay = 10;
-  export let offset = 200;
+  interface Props {
+    value?: number;
+    delay?: number;
+    offset?: number;
+  }
 
-  let i = Math.max(value - offset, 0);
-  let t;
+  let { value = 10, delay = 10, offset = 200 }: Props = $props();
 
-  $: {
+  let i = $state(Math.max(value - offset, 0));
+  let t = $state();
+
+  run(() => {
     if (i === value) {
       if (t) {
         clearInterval(t);
         t = null;
       }
     }
-  }
+  });
 
   onMount(() => {
     t = setInterval(() => {
