@@ -1,6 +1,10 @@
 <script lang="ts">
-  export let isbn: string;
-  export let size = "M";
+  interface Props {
+    isbn: string;
+    size?: string;
+  }
+
+  let { isbn, size = "M" }: Props = $props();
 
   let img_url = `https://covers.openlibrary.org/b/isbn/${isbn}-${size}.jpg`;
   let book_url = `https://openlibrary.org/isbn/${isbn}`;
